requirements.txt
cryptography
flask
python-dotenv
requests
